% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetTables.R
\docType{data}
\name{bbs_cache}
\alias{bbs_cache}
\alias{.bbs_cache}
\title{Custom Environment to store downloaded parquet tibbles.}
\format{
An environment
}
\usage{
.bbs_cache
}
\description{
A custom environment to store remotely read parquet files of
Bioconductor build reports. Once one of the three valid parquet files are
read in, the table is cached in this environment for quick easy reference
either to use ad hocly by the user or to be used within pre-defined queries
provided by this package.
}
\keyword{datasets}
