% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddb.R
\name{current_report_tgz_links}
\alias{current_report_tgz_links}
\title{Get links for all current report.tgz}
\usage{
current_report_tgz_links()
}
\value{
character vector of all current report.tgz links
}
\description{
Get links for all current report.tgz
}
\examples{

current_report_tgz_links()

}
\author{
Sean Davis <seandavi@gmail.com>

current_report_tgz_links()
}
