% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddb.R
\name{example_report_tgz}
\alias{example_report_tgz}
\title{An example report.tgz file for testing}
\usage{
example_report_tgz()
}
\value{
A path to an example report.tgz file
}
\description{
An example report.tgz file for testing
}
\examples{
example_report_tgz()

}
\author{
Sean Davis <seandavi@gmail.com>
}
