% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetTables.R
\name{get_all_bbs_tables}
\alias{get_all_bbs_tables}
\title{Retrieve all available Bioconductor build system parquet files}
\usage{
get_all_bbs_tables(assign_to_global = FALSE)
}
\arguments{
\item{assign_to_global}{logical(1) indicating if downloaded tables should be
assigned to the current global environment for access. Default is FALSE and
only assigned to custom environment through call to get_bbs_table.}
}
\value{
invisible. tables are cached to custom environment unless
assign_to_global is TRUE.
}
\description{
Retrieve all available parquet tables at once.
}
\details{
This function is a quick wrapper around get_bbs_table to download
all available parquet files as tibbles.  By default no value is returned and
the tibbles are stored to the custom environment and can be easily accessed
through a call to get_bbs_table.
}
\examples{
get_all_bbs_tables()

# Now that all are downloaded. Retrieve cached build_summary tibble
build_summary <- get_bbs_table("build_summary")

library(stringr)
library(dplyr)
# find the times the package BiocFileCache failed on linux nebbiolo builders
\dontrun{
build_summary |> filter(package == "BiocFileCache", status \%in\% c("TIMEOUT", "ERROR"), str_starts(node, "nebbiolo")) 
}
}
\seealso{
get_bbs_table
}
\author{
Lori Shepherd
}
