% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetTables.R
\name{get_bbs_table}
\alias{get_bbs_table}
\title{Retrieve a remotely read parquet file of Bioconductor build data}
\usage{
get_bbs_table(tblname = c("build_summary", "info", "propagation_status"))
}
\arguments{
\item{tblname}{A valid parquet file name. Currently available: build_summary,
info, or propagation_status.}
}
\value{
returns a tibble.
}
\description{
Bioconductor Build results are at times cumbersome to work with and
interrogate. Parquet files of condensed reports are provided remotely. These
functions read the remote parquet files using the arrow package and return a
tibble. Once retrieved the table is stored in a custom environment for easy
quick reference either to use ad hocly or within pre-defined queries provided
by this package.
}
\details{
The get_bbs_table returns a tibble but is also saved in a custom
environment for quick easy subsequent reference. There are three tables
available and can be specified in get_bbs_table: build_summary, info, or
propagation_status. build_summary returns a tibble with the results of each
stage of the Bioconductor build process. info returns a tibble with
information regarding the package maintainer and git commit
status. propagation_status returns a tibble with data on if a package
propagated to the community.
}
\examples{
info <- get_bbs_table("info")
info

# all the package Lori Shepherd maintains
info |> filter(Maintainer == "Lori Shepherd") |> distinct(Package)


}
\seealso{
get_all_bbs_tables
}
\author{
Lori Shepherd
}
