% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{get_build_report}
\alias{get_build_report}
\title{Get a report for any given day from the Bioconductor Build System}
\usage{
get_build_report(build_date = Sys.Date(), branch = NULL, builder = NULL)
}
\arguments{
\item{build_date}{Date of report to retrieve (eg. '2025-12-19')}

\item{branch}{a bioconductor branch (eg. 'devel', 'RELEASE_3_22')}

\item{builder}{name of a valid Bioconductor build machine (ie 'nebbiolo1')}
}
\value{
tibble representing a Bioconductor build report for a given day
}
\description{
Get a report for any given day from the Bioconductor Build
System, optionally specifying a git branch or build machine name.
}
\details{
Utilizes data from the build_summary and info parquet files to
retrieve a reconstructed Bioconductor Build Report for any given
day. Optionally a given build machine name or Bioconductor branch may be
specified for further filtering. By default it will use today's date and
retrieve both devel and current release for all Bioconductor build machines.
NOTE: Traditionally Bioconductor splits builds into software,
data-experiment, annotation, workflow, books; this distinction is not
available in these functions and packages regardless of 'type' will be
included.
The returned tibble includes: package name, build node, build stage, package
version number, status of stage, when the run started and ended, the command
the builder ran, branch of bioconductor and relevant git commit information.
}
\examples{
get_build_report("2025-12-29", branch="RELEASE_3_22", builder="nebbiolo2")

}
\seealso{
get_bbs_table
}
\author{
Lori Shepherd
}
