% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddb.R
\name{get_build_summary_table}
\alias{get_build_summary_table}
\title{Extract build_summary table from an extracted report tarball}
\usage{
get_build_summary_table(report_dir)
}
\arguments{
\item{report_dir}{The directory containing the extracted report tarball.
The report_dir should have been created by localize_report_tgz and
should have class 'report_dir' with at least the `md5` attribute.}
}
\value{
A data frame with one row per build process per package
}
\description{
Extract build_summary table from an extracted report tarball
}
\examples{
report_tgz = example_report_tgz()
report_dir = untar_report_tgz(report_tgz)
get_build_summary_table(report_dir)

}
\author{
Sean Davis <seandavi@gmail.com>
}
