% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{get_failing_packages}
\alias{get_failing_packages}
\title{Get list of failing packages for the most recent build}
\usage{
get_failing_packages(branch = NULL, builder = NULL)
}
\arguments{
\item{branch}{a bioconductor branch (eg. 'devel', 'RELEASE_3_22')}

\item{builder}{name of a valid Bioconductor build machine (ie 'nebbiolo1')}
}
\value{
a tibble of failing packages
}
\description{
Get a list of failing (ERROR/TIMEOUT) packages for any given
branch and builder
}
\details{
Utilizes data from the build_summary and info parquet files to list
all packages failing (ERROR/TIMEOUT) in any stage of the process. Optionally
a Bioconductor git branch can be specified and by default will do both the
current devel and release. The report can be further filtered by optionally
specifying a build machine name.
The returned tibble includes: git branch, package name, package version
number, build machine node, collapsed column of stages and statuses.
}
\examples{
get_failing_packages("RELEASE_3_22")
get_failing_packages("RELEASE_3_22", "nebbiolo2")

}
\seealso{
get_bbs_table
}
\author{
Lori Shepherd
}
