% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{get_latest_branches}
\alias{get_latest_branches}
\title{Get currently active Bioconductor git branches}
\usage{
get_latest_branches(infoTbl = NULL)
}
\arguments{
\item{infoTbl}{downloaded paraquet info table retrieved from
get_bbs_table('info') that includes git_branch information.}
}
\value{
character(2) 'devel' and RELEASE_X_Y value eqivalent to current
release.
}
\description{
Get currently active Bioconductor git branches
}
\details{
Get currently active Bioconductor git branches
}
\examples{
get_latest_branches()

}
\seealso{
get_bbs_table
}
