% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{get_package_build_results}
\alias{get_package_build_results}
\title{Get the most recent build result data for a given package on a given branch}
\usage{
get_package_build_results(packagename, branch = "devel")
}
\arguments{
\item{packagename}{a single valid Bioconductor package name}

\item{branch}{a bioconductor branch (eg. 'devel', 'RELEASE_3_22')}
}
\value{
tibble of build status information or NULL if package or branch is not found
}
\description{
Get the build status information for a given package on a specified branch.
}
\details{
Utilizes data from the info and build_summary parquet files to
retrieve the latest build results for a specified package on a given
branch. If no branch is given it defaults to use 'devel'. The data returned
includes the node (builder machine name), stage (install, build, check),
package version number, status of stage, date stage completed, and git
information for the build (git branch, git commit hash, git commit date).
}
\examples{
get_package_build_results("BiocFileCache")
get_package_build_results("BiocFileCache", branch="RELEASE_3_22")

}
\seealso{
get_bbs_table
}
\author{
Lori Shepherd
}
