% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{get_package_release_info}
\alias{get_package_release_info}
\title{Get git information for a given package across all Bioconductor releases}
\usage{
get_package_release_info(packagename)
}
\arguments{
\item{packagename}{a single valid Bioconductor package name}
}
\value{
tibble of release information or NULL if package is not found
}
\description{
Get the package version and git commit data for all releases in Bioconductor.
}
\details{
Utilizes data from the info parquet file to retrieve the equivalent
git information for every available Bioconductor release of a package. This
include the git commit hash and last commit date along with the package version.
}
\examples{
get_package_release_info("BiocFileCache")

}
\seealso{
get_bbs_table
}
\author{
Lori Shepherd
}
