% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddb.R
\name{get_propagation_status_table}
\alias{get_propagation_status_table}
\title{Extract propagation status table from an extracted report tarball}
\usage{
get_propagation_status_table(report_dir)
}
\arguments{
\item{report_dir}{The directory containing the extracted report tarball}
}
\value{
A data frame with one row per package per process
}
\description{
Extract propagation status table from an extracted report tarball
}
\examples{
report_tgz = example_report_tgz()
report_dir = untar_report_tgz(report_tgz)
get_propagation_status_table(report_dir)

}
\author{
Sean Davis <seandavi@gmail.com>
}
