% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddb.R
\name{last_modified_date}
\alias{last_modified_date}
\title{Get the last_modified_date for a Bioconductor URL}
\usage{
last_modified_date(url)
}
\arguments{
\item{url}{A URL that reports the `Last-Modified` header
in the format `\%A, \%d \%B \%Y \%H:\%M:\%S GMT`}
}
\value{
A `POSIXct` object
}
\description{
Makes the assumption that the header name
is `Last-Modified` and that returned value is
parseable by clock::parse_datetime(). Bioconductor
uses the format `\%A, \%d \%B \%Y \%H:\%M:\%S GMT`
}
\examples{
last_modified_date("https://bioconductor.org/checkResults/3.14/bioc-LATEST/report.tgz")

}
\author{
Sean Davis <seandavi@gmail.com>
}
