% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddb.R
\name{localize_report_tgz}
\alias{localize_report_tgz}
\title{Localize a report.tgz file to a local directory}
\usage{
localize_report_tgz(url, local_dir = tempdir())
}
\arguments{
\item{url}{character(1) The URL of the report.tgz file}

\item{local_dir}{character(1) The directory to which the report.tgz file
will be copied (default: tempdir())}
}
\value{
A character(1) with the path to the localized report.tgz file.
}
\description{
This function is used to pull down a report.tgz file from
Bioconductor and save it locally. The file is saved with
a name that is the md5sum of the file, and the file is
saved to the `local_dir`.
}
\examples{
localize_report_tgz("https://bioconductor.org/checkResults/3.14/bioc-LATEST/report.tgz")

}
