% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{package_error_count}
\alias{package_error_count}
\title{Get a count of how many time a package failed on the Bioconductor
Build System}
\usage{
package_error_count(packagename, builder = NULL, branch = NULL)
}
\arguments{
\item{packagename}{a single valid Bioconductor package name}

\item{builder}{name of a valid Bioconductor build machine (ie 'nebbiolo1')}

\item{branch}{a bioconductor branch (eg. 'devel', 'RELEASE_3_22')}
}
\value{
tibble of data including counts of runs and errors or NULL if package
is not found
}
\description{
Get a count of how many time a package failed (ERROR/TIMEOUT) on the
Bioconductor Build System; optionally can specify a specific builder or
branch to filter results.
}
\details{
Utilizes data from the build_summary and info parquet files to
retrieve a count of errors (TIMEOUT/ERROR) for  a specified
package. Optionally a given builder machine name or Bioconductor branch may
be specified to filter request. If no builder is given it will return all. If
no branch is give it will also return all. The data returned
includes the node (builder machine name), package version number, stage
(install, build, check), total number of times that stage/builder tried to
run, total number of times the package failed during that stage, and the
given Bioconductor git branch.
NOTE: since 'devel' is on-going. There will be multiple 'devel' results
across multiple years not just the current devel. See example to filtering to
see most recent.
NOTE: If builder or branch is not found, function will continue and return
unfiltered results.
}
\examples{
package_error_count("BiocFileCache")
package_error_count("BiocFileCache", branch="RELEASE_3_22")
package_error_count("BiocFileCache", builder = "nebbiolo2", branch="RELEASE_3_22")

library(dplyr)
## devel will have more than just current devel
devErrors <- package_error_count("BiocFileCache", branch="devel")
## can use package version number to filter to current devel
devErrors |> filter(version == max(version))

}
\seealso{
get_bbs_table
}
\author{
Lori Shepherd
}
