% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{package_failures_over_time}
\alias{package_failures_over_time}
\alias{package_failure_over_time}
\title{Get rough timeline overview of package failing events}
\usage{
package_failures_over_time(packagename, builder, failure_cluster_hours = 72)
}
\arguments{
\item{packagename}{a single valid Bioconductor package name}

\item{builder}{name of a valid Bioconductor build machine (ie 'nebbiolo1')}

\item{failure_cluster_hours}{numeric indicating the number of hours to treat
failures as a single event}
}
\value{
tibble of failure event information or NULL if package or builder is not found
}
\description{
Get rough timeline overview of a package failing events on a
specific builder
}
\details{
This function attempts to give an overview of how long a package has
been failing on a given builder. The time to treat failures as a single event
is distinguished by consecutive days and by the failure_cluster_hours
argument. Currently there is a different cadence of builds for release
compared to devel so this allows the flexibility to define the timeline.
Let's use the following example for interpretation of the result table:
```
  version episode       first_failure        last_failure n_failures
2   3.1.0       3 2025-12-15 02:57:57 2025-12-16 21:29:20          3
3   3.1.0       2 2025-12-13 02:32:33 2025-12-13 02:32:33          1
4   3.1.0       1 2025-12-08 21:30:14 2025-12-11 21:27:29          4
             stages statuses
2 buildsrc, checksrc    ERROR
3           checksrc    ERROR
4           buildsrc    ERROR
```
The package version being built on the build system for this package is
3.1.0. The episode refers to the number of failing episodic events. If we
look at the episode 1, the package started failing on December 8 and
continued to fail until December 11. During that time period there were 4
number of failures and they occurred during the build stage. The time between
the last failure of the first episode and the second episode is from December
11 to December 13; we can infer then that the package built cleanly on
December 12. During episode 2 the package failed solely on December 13 but
during the check stage. Episode 3 began on Dec 15 and lasted until December
16 and had 3 failures but they occurred in either the build or check stage.

This hopefully can help determine if package failures are intermittent and
how long a package is failing on a given builder.
}
\examples{
package_failures_over_time("BiocFileCache", "nebbiolo1", 24)

}
\seealso{
get_bbs_table
}
\author{
Lori Shepherd
}
