% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{process_all_new_reports}
\alias{process_all_new_reports}
\title{Get and process all new build reports}
\usage{
process_all_new_reports(reportdb_file, local_dir)
}
\arguments{
\item{reportdb_file}{character(1) Filename of the reportdb file with last modification dates.
The file does not need to exist. If it does, it will be read and updated.
If it does not, it will be created.}

\item{local_dir}{The directory to which the report.tgz files will be copied,
untarred, and processed.}
}
\description{
Get and process all new build reports
}
