% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_db.R
\name{report_last_mod_db}
\alias{report_last_mod_db}
\title{read a little file-based database of last-modified dates for reports}
\usage{
report_last_mod_db(reportdb_file)
}
\arguments{
\item{reportdb_file}{the file to read the database from}
}
\value{
a tibble with columns url and old_last_modified
}
\description{
If the file does not exist, the function returns an empty tibble
with the expected columns.
}
