% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddb.R
\name{reports_last_mod_table}
\alias{reports_last_mod_table}
\title{Get the last_modified_date for all current report.tgz}
\usage{
reports_last_mod_table()
}
\value{
A data frame with one row per report.tgz file
}
\description{
Get the last_modified_date for all current report.tgz
}
\examples{
res = reports_last_mod_table()
res

}
\author{
Sean Davis <seandavi@gmail.com>
}
