% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{tables_from_report_dir}
\alias{tables_from_report_dir}
\title{Create tables for a report directory}
\usage{
tables_from_report_dir(report_dir)
}
\arguments{
\item{report_dir}{character(1) The directory containing the report files.}
}
\value{
A list of data.frames with the tables from the report directory.
}
\description{
This function creates a set of tables from a report directory.
}
