% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddb.R
\name{untar_report_tgz}
\alias{untar_report_tgz}
\title{Untar a report.tgz file and return a simple `report_dir` object}
\usage{
untar_report_tgz(local_report_tgz, local_dir = tempdir())
}
\arguments{
\item{local_report_tgz}{character(1) The path to the report.tgz file}

\item{local_dir}{character(1) The directory to which the report.tgz file
will be untarred (default: tempdir())}
}
\value{
A character(1) with the path to the untarred report directory
  with class `report_dir` and with attributes `md5`, `url`, and `local_report_tgz`.
}
\description{
This function is used to untar a report.tgz file and return a simple
`report_dir` object. The `report_dir` object is a character(1) with
class `report_dir` and with attributes `md5`, `url`, and `local_report_tgz`.
}
\examples{
report_tgz = example_report_tgz()
report_dir = untar_report_tgz(report_tgz)

head(list.files(report_dir, recursive = TRUE))

}
