% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{write_tables_to_files}
\alias{write_tables_to_files}
\title{Write tables to files based on md5 naming}
\usage{
write_tables_to_files(table_list, out_dir, md5)
}
\arguments{
\item{table_list}{A list of data.frames to be written to files.}

\item{out_dir}{The directory to which the files will be written.}

\item{md5}{The md5 hash of the report directory. Will be used
as a prefix for the files.}
}
\description{
Write tables to files based on md5 naming
}
